Fix selection and move annotation drawing into annotation set
