+++ projects/media-viewer/src/lib/annotations/annotation-set/annotation-set.component.spec.ts
@@ -10,6 +10,7 @@ import { AnnotationComponent } from './annotation/annotation.component';
 import { AnnotationApiService } from '../annotation-api.service';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { of } from 'rxjs';
+import { ElementRef } from '@angular/core';
 
 describe('AnnotationSetComponent', () => {
   let component: AnnotationSetComponent;
@@ -17,8 +18,8 @@ describe('AnnotationSetComponent', () => {
 
   const api = new AnnotationApiService({}  as any);
 
-  beforeEach(async(() => {
-    return TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       declarations: [
         AnnotationSetComponent,
         AnnotationComponent,
@@ -34,11 +35,8 @@ describe('AnnotationSetComponent', () => {
       providers: [
         { provide: AnnotationApiService, useValue: api }
       ]
-    })
-      .compileComponents();
-  }));
+    }).compileComponents();
 
-  beforeEach(() => {
     fixture = TestBed.createComponent(AnnotationSetComponent);
     component = fixture.componentInstance;
     component.annotationSet = annotationSet;
@@ -70,6 +68,34 @@ describe('AnnotationSetComponent', () => {
   it('deselects an annotation', () => {
     component.onAnnotationSelected(false, 1);
 
+    expect(component.selected).toEqual(-1);
+  });
+
+  it('starts drawing on mousedown', () => {
+    component.newRectangle = new ElementRef(document.createElement('div'));
+    component.drawMode = true;
+    component.onMouseDown({ pageY: 10, pageX: 10 } as MouseEvent);
+
+    expect(component.newRectangle.nativeElement.style.left).toEqual('10px');
+    expect(component.newRectangle.nativeElement.style.top).toEqual('10px');
+  });
+
+  it('finishes drawing on mouseup', () => {
+    const annotation = Object.assign({}, annotationSet.annotations[0]);
+    annotation.id = 'new';
+
+    const spy = spyOn(api, 'postAnnotation').and.returnValues(of(annotation));
+
+    component.newRectangle = new ElementRef(document.createElement('div'));
+    component.drawMode = true;
+    component.selected = 1;
+    component.onMouseDown({ pageY: 10, pageX: 10 } as MouseEvent);
+    component.onMouseMove({ pageY: 100, pageX: 100 } as MouseEvent);
+    component.onMouseUp();
+
+    expect(spy).toHaveBeenCalled();
+
+    expect(component.annotationSet.annotations[component.annotationSet.annotations.length - 1].id).toEqual('new');
     expect(component.selected).toEqual(-1);
   });
