+++ projects/media-viewer/src/lib/viewers/image-viewer/image-viewer.component.scss
@@ -40,15 +40,3 @@ img {
   -o-transform-origin: left top;
   transform-origin: left top;
 }
-
-.newRectangle {
-  display: none;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 0;
-  height: 0;
-  z-index: 10;
-  opacity: 0.5;
-  background-color: yellow;
-}
